# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Laravel",
        "contact": {
            "name": "hello@example.com",
            "email": "hello@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/medications/store": {
            "post": {
                "tags": [
                    "Medications"
                ],
                "summary": "Store new Medication",
                "description": "Store new Medication",
                "operationId": "medication_store",
                "requestBody": {
                    "description": "new Medication",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicationCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/medications/show/{id}": {
            "get": {
                "tags": [
                    "Medications"
                ],
                "summary": "Select existing Medication",
                "description": "Select existing Medication",
                "operationId": "medication_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Medication id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/medications/filter": {
            "get": {
                "tags": [
                    "Medications"
                ],
                "summary": "Filter existing Medications",
                "description": "Filter existing Medications",
                "operationId": "medication_filter",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Page Limit",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 0
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/medications/update/{id}": {
            "put": {
                "tags": [
                    "Medications"
                ],
                "summary": "Update existing Medication",
                "description": "Update existing Medication",
                "operationId": "medication_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Medication id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update existing Medication",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicationUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/medications/flag-as-deleted/{id}": {
            "delete": {
                "tags": [
                    "Medications"
                ],
                "summary": "Soft Delete existing Medication",
                "description": "Soft Delete existing Medication",
                "operationId": "medication_flagAsDeleted",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Medication id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/medications/permanently-delete-records/{id}": {
            "delete": {
                "tags": [
                    "Medications"
                ],
                "summary": "Force Delete existing Medication",
                "description": "Force Delete existing Medication",
                "operationId": "medication_destroyPermanently",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Medication id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MedicationCreateRequest": {
                "title": "Medication Create request",
                "description": "Medication Create request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Medication Name",
                        "type": "string",
                        "example": "Medication 01"
                    },
                    "description": {
                        "title": "description",
                        "description": "Medication Description",
                        "type": "string",
                        "example": "Description"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "Medication Quantity",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "MedicationUpdateRequest": {
                "title": "Medication Update request",
                "description": "Medication Update request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Medication Name",
                        "type": "string",
                        "example": "Medication 01"
                    },
                    "description": {
                        "title": "description",
                        "description": "Medication Description",
                        "type": "string",
                        "example": "Description"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "Medication Quantity",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "JsonResponseResource": {
                "title": "JsonResponseResource",
                "description": "Json Response Resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status",
                        "type": "string"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data",
                        "nullable": true
                    },
                    "errors": {
                        "title": "Errors",
                        "description": "Errors",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonResponseResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}